Parser and epub remastering
Change Chrome Browser to HttpClient
Chande epub engine
Bug fixes

+++ NovelParserBLL/Extensions/SortedListChaptersExtension.cs
@@ -4,7 +4,7 @@ namespace NovelParserBLL.Extensions
 {
     internal static class SortedListChaptersExtension
     {
-        public static List<Chapter> ForLoad(this SortedList<int, Chapter> chapters, bool includeImages)
+        public static List<Chapter> ForLoad(this SortedList<float, Chapter> chapters, bool includeImages)
         {
             return chapters.Select(v => v.Value).Where(ch => string.IsNullOrEmpty(ch.Content) || (!ch.ImagesLoaded || !ch.Images.Exists()) && includeImages).ToList();
         }
